Refactor codex and inventory helpers into dedicated modules
